EDMT-397 refactor : api 명세 변경에 따른 auth mock api status 제거 및 code 수정
